Moved plot._utils to plot._vispy_utils.
